fix: for addChoices, InputChoice from client doesn't match server
Fixed by just not properly validating the InputChoice argument
that was received from client, which is not a real fix at all.

The problematic factor should be the DNDVotes field? In other words,
DNDVotes are probably being sent by the client to the server in the
wrong format, leading to a argument validation error.

The server is probably storing nulls for the DNDVotes field, even
though in the request the client intended to include a DNDVotes.